factor out fallacy detection into its own module
