+++ app/components/building-blocks/post-info-bar.tsx
@@ -1,6 +1,6 @@
 import moment from 'moment'
 import { useState } from 'react'
-import { type FallacyList } from '#app/repositories/fallacy-detection.ts'
+import { type FallacyList } from '#app/modules/fallacies/fallacy-types.ts'
 import { type PostState, type Post } from '#app/types/api-types.ts'
 
 export function PostInfoBar({
