+++ app/routes/_actions+/reply.tsx
@@ -2,10 +2,8 @@ import { type ActionFunctionArgs } from '@remix-run/node'
 import invariant from 'tiny-invariant'
 import { z } from 'zod'
 import { db } from '#app/db.ts'
-import {
-	fallacyDetection,
-	storeFallacies,
-} from '#app/repositories/fallacy-detection.ts'
+import { fallacyDetection } from '#app/modules/fallacies/fallacy-detection-client.ts'
+import { storeFallacies } from '#app/modules/fallacies/fallacy-repository.ts'
 import { createPost } from '#app/repositories/post.ts'
 import { getCommentTreeState, getReplyTree } from '#app/repositories/ranking.ts'
 import { requireUserId } from '#app/utils/auth.server.ts'
