+++ app/routes/_actions+/submit-artefact.tsx
@@ -3,7 +3,7 @@ import { z } from 'zod'
 import { db } from '#app/db.ts'
 import { extractClaims } from '#app/modules/claim-extraction/claim-extraction-client.ts'
 import { getOrCreateArtefact } from '#app/repositories/artefact.ts'
-import { getOrDetectQuoteFallacies } from '#app/repositories/fallacy-detection.ts'
+import { getOrDetectQuoteFallacies } from '#app/repositories/quote-fallacy.ts'
 import { getOrCreateQuote } from '#app/repositories/quote.ts'
 
 const artefactDtoSchema = z.object({
